[FIX] Re-arranges directory structure, adds licences, contributors
