+++ help_popup/models/__init__.py
@@ -2,4 +2,4 @@
 # Copyright (C) 2015-TODAY Akretion (<http://www.akretion.com>).
 # © 2017 Therp BV <http://therp.nl>
 # License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl.html).
-from . import models
+from . import model
